alerts new followers with alert( follower )
is the follower on the session list already? if so, continue to the
next, or else add it to the session list and alert it. got to create a
pretty animation next, instead of the browser alert. also set up a way
to keep a perpetual save of the session list: Firebase?